+++ spec/groupdocs/signature/list_spec.rb
@@ -12,7 +12,7 @@ describe GroupDocs::Signature::List do
     it 'accepts access credentials hash' do
       lambda do
         described_class.get!(:client_id => 'client_id', :private_key => 'private_key')
-      end.should_not raise_error(ArgumentError)
+      end.should_not raise_error()
     end
 
     it 'returns array of GroupDocs::Signature::List objects' do
@@ -58,7 +58,7 @@ describe GroupDocs::Signature::List do
     it 'accepts access credentials hash' do
       lambda do
         subject.add!(:client_id => 'client_id', :private_key => 'private_key')
-      end.should_not raise_error(ArgumentError)
+      end.should_not raise_error()
     end
 
     it 'uses hashed version of self as request body' do
@@ -81,7 +81,7 @@ describe GroupDocs::Signature::List do
     it 'accepts access credentials hash' do
       lambda do
         subject.delete!(:client_id => 'client_id', :private_key => 'private_key')
-      end.should_not raise_error(ArgumentError)
+      end.should_not raise_error()
     end
   end
 end