+++ spec/groupdocs/storage/package_spec.rb
@@ -22,13 +22,13 @@ describe GroupDocs::Storage::Package do
   describe '#create!' do
     before(:each) do
       mock_api_server(load_json('package_create'))
-      subject.objects = [stub(:name => 'object 1', :path => 'path 1')]
+      subject.objects = [double(:name => 'object 1', :path => 'path 1')]
     end
 
     it 'accepts access credentials hash' do
       lambda do
         subject.create!(:client_id => 'client_id', :private_key => 'private_key')
-      end.should_not raise_error(ArgumentError)
+      end.should_not raise_error()
     end
 
     it 'returns URL for package downloading' do