+++ spec/support/shared_examples/signature/shared/recipient_methods.rb
@@ -8,11 +8,11 @@ shared_examples_for GroupDocs::Signature::RecipientMethods do
     it 'accepts access credentials hash' do
       lambda do
         subject.recipients!({}, :client_id => 'client_id', :private_key => 'private_key')
-      end.should_not raise_error(ArgumentError)
+      end.should_not raise_error()
     end
 
     it 'can be public' do
-      lambda { subject.recipients!(:public => true) }.should_not raise_error(ArgumentError)
+      lambda { subject.recipients!(:public => true) }.should_not raise_error()
     end
 
     it 'returns array of GroupDocs::Signature::Recipient objects' do
@@ -36,7 +36,7 @@ shared_examples_for GroupDocs::Signature::RecipientMethods do
     it 'accepts access credentials hash' do
       lambda do
         subject.remove_recipient!(recipient, :client_id => 'client_id', :private_key => 'private_key')
-      end.should_not raise_error(ArgumentError)
+      end.should_not raise_error()
     end
 
     it 'raises error if recipient is not GroupDocs::Signature::Recipient object' do